error hantering i settings tillagd

+++ MysteryLocation/Model/User.cs
@@ -17,7 +17,7 @@ namespace MysteryLocation.Model
 
 
         public int category;
-        public int distance;
+        public double distance;
 
         private Post tracking { get; set; }
         private int tracker;
@@ -45,7 +45,7 @@ namespace MysteryLocation.Model
             category = cat;
             Console.WriteLine("The user category has been changed to " + category.ToString());
         }
-        public void setDistance(int dist)
+        public void setDistance(double dist)
         {
             distance = dist;
             Console.WriteLine("The user distance has been changed to " + distance.ToString());
